Do not write files in user system
This stops sysinfo from writing files in the running system without
telling the user about it.

A system with no machine-id is a completely valid system and can be like
that at different times, like inside a container, on a first boot, in
immutable systems, etc...

We should not write files in teh user system without informing them and
not having a machine-id is a valid state.

Signed-off-by: Itxaka <itxaka@kairos.io>